refactor environments and utils
