+++ backend/data_tools/Pipfile
@@ -19,6 +19,7 @@ alembic-postgresql-enum = "==1.3.0"
 azure-storage-blob = "==12.22.0"
 azure-identity = "==1.17.1"
 azure-keyvault-secrets = "==4.8.0"
+loguru = "*"
 
 [dev-packages]
 nox = "==2024.4.15"
