+++ backend/data_tools/environment/dev.py
@@ -1,4 +1,4 @@
-from data_tools.environment.common import DataToolsConfig
+from data_tools.environment.types import DataToolsConfig
 
 
 class DevConfig(DataToolsConfig):
@@ -7,10 +7,6 @@ class DevConfig(DataToolsConfig):
         return (
             "postgresql://ops:ops@localhost:5432/postgres"  # pragma: allowlist secret
         )
-
-    @property
-    def opre_excel_connection_string(self) -> str:
-        return ""
 
     @property
     def verbosity(self) -> bool:
