+++ backend/data_tools/environment/local.py
@@ -1,14 +1,10 @@
-from data_tools.environment.common import DataToolsConfig
+from data_tools.environment.types import DataToolsConfig
 
 
 class LocalConfig(DataToolsConfig):
     @property
     def db_connection_string(self) -> str:
         return "postgresql://ops:ops@db:5432/postgres"  # pragma: allowlist secret
-
-    @property
-    def opre_excel_connection_string(self) -> str:
-        return ""
 
     @property
     def verbosity(self) -> bool:
