+++ backend/data_tools/environment/types.py
@@ -8,10 +8,6 @@ class DataToolsConfig(Protocol):
     def db_connection_string(self) -> str:
         ...
 
-    @property
-    @abstractmethod
-    def opre_excel_connection_string(self) -> str:
-        ...
 
     @property
     @abstractmethod
