+++ backend/data_tools/src/etl_data_from_excel/import_project_data_from_excel.py
@@ -1,43 +1,6 @@
 import os
-from typing import Tuple
 
-import sqlalchemy.engine
-from data_tools.environment.local_migration import LocalMigrationConfig
-from environment.cloudgov import CloudGovConfig
-from environment.common import DataToolsConfig
-from environment.dev import DevConfig
-from environment.local import LocalConfig
-from environment.test import TestConfig
-from sqlalchemy import MetaData, create_engine
-
-
-def init_db(
-    config: DataToolsConfig,
-) -> Tuple[sqlalchemy.engine.Engine, sqlalchemy.MetaData]:
-    engine = create_engine(
-        config.db_connection_string, echo=config.verbosity, future=True
-    )
-    metadata_obj = MetaData()
-    metadata_obj.reflect(bind=engine)
-
-    return engine, metadata_obj
-
-
-def get_config(environment_name: str) -> DataToolsConfig:
-    config: DataToolsConfig
-    match environment_name:
-        case "cloudgov":
-            config = CloudGovConfig()
-        case "local":
-            config = LocalConfig()
-        case "local-migration":
-            config = LocalMigrationConfig()
-        case "test":
-            config = TestConfig()
-        case _:
-            config = DevConfig()
-    return config
-
+from common.utils import get_config, init_db
 
 if __name__ == "__main__":
     script_env = os.getenv("ENV")
