+++ backend/ops_api/pytest.ini
@@ -2,4 +2,4 @@
 pythonpath = . ..
 testpaths = tests
 bdd_features_base_dir = tests/ops/features/
-# addopts = --cov
+addopts = --cov
